+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/ui/serviceconnection/SplitTunneling.kt
@@ -7,7 +7,8 @@ import net.mullvad.mullvadvpn.ipc.EventDispatcher
 import net.mullvad.mullvadvpn.ipc.Request
 
 class SplitTunneling(private val connection: Messenger, eventDispatcher: EventDispatcher) {
-    private var excludedApps: Set<String> = emptySet()
+    private var _excludedApps by
+        observable(emptySet<String>()) { _, _, apps -> excludedAppsChange.invoke(apps) }
 
     var enabled by
         observable(false) { _, wasEnabled, isEnabled ->
@@ -16,18 +17,22 @@ class SplitTunneling(private val connection: Messenger, eventDispatcher: EventDi
             }
         }
 
+    var excludedAppsChange: (apps: Set<String>) -> Unit = {}
+        set(value) {
+            field = value
+            synchronized(this) { value.invoke(_excludedApps) }
+        }
+
     init {
         eventDispatcher.registerHandler(Event.SplitTunnelingUpdate::class) { event ->
             if (event.excludedApps != null) {
                 enabled = true
-                excludedApps = event.excludedApps.toSet()
+                _excludedApps = event.excludedApps.toSet()
             } else {
                 enabled = false
             }
         }
     }
-
-    fun isAppExcluded(appPackageName: String): Boolean = excludedApps.contains(appPackageName)
 
     fun excludeApp(appPackageName: String) =
         connection.send(Request.ExcludeApp(appPackageName).message)