+++ CHANGELOG.md
@@ -49,6 +49,7 @@ Line wrap the file at 100 chars.                                              Th
 - Clarify some of the error messages showed in notifications.
 - Change and add text in Content blockers informational dialog.
 - Update advanced settings scrollbar appearance.
+- Migrate split tunneling view to compose
 
 ### Fixed
 #### Android